ekf landmarks runs but does not work

+++ src/models/motion.rs
@@ -12,8 +12,8 @@ pub trait MotionModel<T: RealField, const S: usize, const Z: usize, const U: usi
 
 pub struct Velocity {}
 
-impl MotionModel<f32, 3, 2, 2> for Velocity {
-    fn prediction(&self, x: &Vector3<f32>, u: &Vector2<f32>, dt: f32) -> Vector3<f32> {
+impl MotionModel<f64, 3, 2, 2> for Velocity {
+    fn prediction(&self, x: &Vector3<f64>, u: &Vector2<f64>, dt: f64) -> Vector3<f64> {
         //state
         let theta = x[2];
         //control
@@ -28,14 +28,14 @@ impl MotionModel<f32, 3, 2, 2> for Velocity {
     }
 
     #[allow(clippy::deprecated_cfg_attr)]
-    fn jacobian_wrt_state(&self, x: &Vector3<f32>, u: &Vector2<f32>, dt: f32) -> Matrix3<f32> {
+    fn jacobian_wrt_state(&self, x: &Vector3<f64>, u: &Vector2<f64>, dt: f64) -> Matrix3<f64> {
         //state
         let theta = x[2];
         //control
         let v = u[0];
         let w = u[1];
         #[cfg_attr(rustfmt, rustfmt_skip)]
-        Matrix3::<f32>::new(
+        Matrix3::<f64>::new(
             1., 0., v / w * (-theta.cos() + (theta + w * dt).cos()),
             0., 1., v / w * (-theta.sin() + (theta + w * dt).sin()),
             0., 0., 1.
@@ -43,7 +43,7 @@ impl MotionModel<f32, 3, 2, 2> for Velocity {
     }
 
     #[allow(clippy::deprecated_cfg_attr)]
-    fn jacobian_wrt_input(&self, x: &Vector3<f32>, u: &Vector2<f32>, dt: f32) -> Matrix3x2<f32> {
+    fn jacobian_wrt_input(&self, x: &Vector3<f64>, u: &Vector2<f64>, dt: f64) -> Matrix3x2<f64> {
         //state
         let theta = x[2];
         //control
@@ -56,7 +56,7 @@ impl MotionModel<f32, 3, 2, 2> for Velocity {
         let costdt = (theta + w * dt).cos();
         let w2 = w * w;
         #[cfg_attr(rustfmt, rustfmt_skip)]
-        Matrix3x2::<f32>::new(
+        Matrix3x2::<f64>::new(
             (-sint + sintdt) / w, v  * ((sint - sintdt) / w2 + costdt * dt / w),
             (cost - costdt) / w, v  * (-(cost - costdt) / w2 + sintdt * dt / w),
             0., 1.
@@ -85,8 +85,8 @@ impl MotionModel<f32, 3, 2, 2> for Velocity {
 /// dy/dv = dt * sin(yaw)
 pub struct SimpleProblemMotionModel {}
 
-impl MotionModel<f32, 4, 2, 2> for SimpleProblemMotionModel {
-    fn prediction(&self, x: &Vector4<f32>, u: &Vector2<f32>, dt: f32) -> Vector4<f32> {
+impl MotionModel<f64, 4, 2, 2> for SimpleProblemMotionModel {
+    fn prediction(&self, x: &Vector4<f64>, u: &Vector2<f64>, dt: f64) -> Vector4<f64> {
         let yaw = x[2];
         let v = x[3];
         Vector4::new(
@@ -98,18 +98,18 @@ impl MotionModel<f32, 4, 2, 2> for SimpleProblemMotionModel {
     }
 
     #[allow(clippy::deprecated_cfg_attr)]
-    fn jacobian_wrt_state(&self, x: &Vector4<f32>, u: &Vector2<f32>, dt: f32) -> Matrix4<f32> {
+    fn jacobian_wrt_state(&self, x: &Vector4<f64>, u: &Vector2<f64>, dt: f64) -> Matrix4<f64> {
         let yaw = x[2];
         let v = u[0];
         #[cfg_attr(rustfmt, rustfmt_skip)]
-            Matrix4::<f32>::new(
+            Matrix4::<f64>::new(
                 1., 0., -dt * v * (yaw).sin(), dt * (yaw).cos(),
                 0., 1., dt * v * (yaw).cos(), dt * (yaw).sin(),
                 0., 0., 1., 0.,
                 0., 0., 0., 0.,
             )
     }
-    fn jacobian_wrt_input(&self, _x: &Vector4<f32>, _u: &Vector2<f32>, _dt: f32) -> Matrix4x2<f32> {
+    fn jacobian_wrt_input(&self, _x: &Vector4<f64>, _u: &Vector2<f64>, _dt: f64) -> Matrix4x2<f64> {
         unimplemented!()
     }
 }
